return filtersArr to filmsContainer for resuablity in cinema page
